statement frequency - api integration
