+++ auth-web/src/models/statement.ts
@@ -1,9 +1,9 @@
 export interface StatementListItem {
-  createdOn: string
+  createdOn?: string
   frequency: string
-  fromDate: string
-  id: number
-  toDate: string
+  fromDate?: string
+  id?: number
+  toDate?: string
 }
 
 export interface StatementFilterParams {
